refactor: extract renderMarkdownToTerminal god function into concern siblings (#1081)
Split the 405-line `renderMarkdownToTerminal` in `src/cli/formatter.ts`
into three concern-scoped siblings:

- `formatter.table.ts` (216 LOC) — column-width water-fill squeeze,
  cell wrapping, border drawing
- `formatter.list.ts` (131 LOC) — ordered/unordered/task lists,
  checkbox glyphs, hanging indent
- `formatter.block.ts` (109 LOC) — blockquote gutter prefix,
  code block syntax highlighting

The original file shrinks from 529 → 198 LOC (now under the 350-line
file ceiling) and the function itself from 405 → 82 LOC (now under the
200-line function ceiling). Both baselines updated accordingly.

All extracted helpers take explicit parameters — no closures over
enclosing locals. The public API of `formatter.ts` is unchanged:
`renderCardLine`, `renderMarkdownToTerminal`, and `renderInlineTokens`
remain at the same import path with identical signatures.
