+++ .filesize-baseline.json
@@ -57,7 +57,7 @@
     "src/agent/tools/subagent/foreground-promotion.ts": { "loc": 452, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/subagent/input-parse.ts": { "loc": 533, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/trace/events.ts": { "loc": 594, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/agent/trace/receipt.ts": { "loc": 538, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/agent/trace/receipt.ts": { "loc": 528, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/trace/types.ts": { "loc": 964, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/trace/writer.ts": { "loc": 530, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/types/config-types.ts": { "loc": 795, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
@@ -95,9 +95,8 @@
     "src/cli/cup-frame-renderer.ts": { "loc": 373, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/elicitation/agent-question.ts": { "loc": 532, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/elicitation/form-mode.ts": { "loc": 413, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/cli/formatter.ts": { "loc": 529, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/input-core.ts": { "loc": 386, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/cli/input/history.ts": { "loc": 368, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/cli/input/history.ts": { "loc": 353, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/input/input-surface.ts": { "loc": 688, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/input/reader.ts": { "loc": 900, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/markdown-stream.ts": { "loc": 408, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
@@ -106,6 +105,7 @@
     "src/cli/slash/commands/info.ts": { "loc": 548, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/slash/marketplace-browse.ts": { "loc": 367, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/slash/plugin-skills/listing.ts": { "loc": 410, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/cli/slash/types.ts": { "loc": 352, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/status-line.ts": { "loc": 621, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/terminal-compositor.frame.ts": { "loc": 370, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/terminal-compositor.input-dispatch.ts": { "loc": 1153, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
