+++ .funcsize-baseline.json
@@ -2,7 +2,7 @@
   "limit": 200,
   "entries": {
     "src/agent/afk-mode-gate.ts::createAfkModeGate": { "loc": 253, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/agent/default-hook-registry.ts::createDefaultHookRegistry": { "loc": 224, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/default-hook-registry.ts::createDefaultHookRegistry": { "loc": 222, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/hooks/command-executor.ts::executeCommand": { "loc": 237, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/anthropic-direct/loop.ts::runTurn": { "loc": 245, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/anthropic-direct/loop/stream-consumer.ts::consumeRoundStream": { "loc": 279, "reason": "legacy: predates the function-size gate; pending helper extraction" },
@@ -15,10 +15,11 @@
     "src/agent/shell-jobs/streamer.ts::startShell": { "loc": 290, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/subagent.ts::SubagentManager.forkSubagent": { "loc": 496, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/subagent/handle.ts::SubagentHandleImpl.streamToFinalMessage": { "loc": 232, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/agent/tools/compose-executor.ts::ComposeExecutor.execute": { "loc": 273, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/tools/compose-executor.ts::ComposeExecutor.execute": { "loc": 329, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/tools/dispatcher.ts::SessionToolDispatcher.executeBatch": { "loc": 201, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/tools/handlers/bash.ts::createBashHandler": { "loc": 318, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/tools/handlers/grep.ts::createGrepHandler": { "loc": 237, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/agent/tools/subagent-executor.ts::SubagentExecutor.executeOnce": { "loc": 384, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/tools/subagent-executor.ts::SubagentExecutor.executeOnce": { "loc": 399, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/tools/subagent/child-config.ts::buildChildConfig": { "loc": 345, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/tools/subagent/foreground-promotion.ts::runForegroundWithPromotion": { "loc": 334, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/tools/subagent/input-parse.ts::parseAgentInput": { "loc": 386, "reason": "legacy: predates the function-size gate; pending helper extraction" },
@@ -44,7 +45,6 @@
     "src/cli/commands/trace.ts::renderEvent": { "loc": 285, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/config/json-tier-parse.ts::parseJsonConfigFile": { "loc": 234, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/elicitation/agent-question.ts::renderAgentQuestion": { "loc": 400, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/formatter.ts::renderMarkdownToTerminal": { "loc": 405, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/input/reader.ts::readWithAutocompleteTty": { "loc": 856, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.frame-preserve.ts::preserveRowsBeforeFrameRender": { "loc": 209, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.input-dispatch.ts::handleEnter": { "loc": 259, "reason": "legacy: predates the function-size gate; pending helper extraction" },
